refactor: 급식과 시간표 날짜 계산 로직 분리
