+++ src/features/wake-up-music/model/useWakeUpMusic.ts
@@ -11,14 +11,14 @@ import {
   dormitoryQueries,
   dormitoryMutations,
 } from "@/entities/dormitory/api/dormitoryQueries";
-import { formatDate } from "@/features/wake-up-music/lib/formatDate";
+import { formatDateParam } from "@/shared/lib/date";
 
 export function useWakeUpMusic() {
   const queryClient = useQueryClient();
   const [urlInput, setUrlInput] = useState("");
   const [selectedDate, setSelectedDate] = useState<Date>(new Date());
 
-  const selectedDateString = formatDate(selectedDate);
+  const selectedDateString = formatDateParam(selectedDate);
   const musicQuery = dormitoryQueries.music(selectedDateString);
 
   const { data: songs = [] } = useQuery(musicQuery);
