+++ src/shared/lib/date.ts
@@ -23,3 +23,13 @@ export function formatDateParam(
 
   return `${formattedDate} ${timeParts.join(":")}`;
 }
+
+export const DAYS = ["일", "월", "화", "수", "목", "금", "토"];
+
+export function formatDisplayDate(date: Date): string {
+  const yy = String(date.getFullYear()).slice(2);
+  const mm = String(date.getMonth() + 1).padStart(2, "0");
+  const dd = String(date.getDate()).padStart(2, "0");
+  const day = DAYS[date.getDay()];
+  return `${yy}.${mm}.${dd} (${day})`;
+}
